Split approot into multiple files
It was beginning to get annoying/confusing with both the main
calculation and summary functions in the same file, since they
were named kind of similarly.

Though this probably could have been done without the moves, the
process of moving and updating the pub/private annotations on each
function forced more granular wasm/non-wasm configs on everything
which sort of helps self-explain what is used where.

approot.rs itself is now fairly small, and it self-explains what
functions are actually intended to be used where (or are suppoesd
to be re-exported into app::).
